backend/pyproject.toml
[project]
name = "baynext-api"
version = "0.1.1"
description = "Baynext API for managing datasets, jobs, and pipelines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "typer>=0.13.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
]

[tool.ruff.lint.isort]
known-local-folder = ["app", "tests"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
env = [
    "ML_API_SECRET_API_KEY=ml_api_secret_api_key",
    "DATABASE_URL=postgresql://user:password@host:80/dbname",
    "BLOB_READ_WRITE_TOKEN=blob_read_write_token",
    "AUTH_SECRET=auth_secret",
]
